+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+### 1.8.2
+- WIN: Add popup.Behaviour.REPLACE Issue #171
+
 ### 1.8.1
-- WIN: add BackgroundColor to HwndSource [Issue #168](https://github.com/tuarua/WebViewANE/issues/168)
+- WIN: add BackgroundColor to HwndSource Issue #168
 
 ### 1.8.0
 - OSX/iOS: Updated to FreSwift 2.5.0
@@ -9,7 +12,7 @@
 
 ### 1.7.0
 - WIN: Updated to CefSharp 65.0.0 pre-rel
-- WIN: UTF-8 bug [Issue #155](https://github.com/tuarua/WebViewANE/issues/155) 
+- WIN: UTF-8 bug Issue #155
 - WIN: Remove MSVC2013 as a dependency
 - WIN: Updated to FreSharp 1.6.0
 - WIN: Added basic support for Modern WebView (Edge)